requirements.txt

selenium
webdriver-manager
pytest
pytest-html
python-dotenv
